Fixing rounding problems in piping. WTI-2875

+++ Riskeer/MacroStabilityInwards/src/Riskeer.MacroStabilityInwards.IO/Factories/PersistableStochasticParameterFactory.cs
@@ -58,7 +58,7 @@ namespace Riskeer.MacroStabilityInwards.IO.Factories
 
         private static double GetStandardDeviation(IVariationCoefficientDistribution distribution)
         {
-            return distribution.Mean * distribution.CoefficientOfVariation;
+            return (double) distribution.Mean * (double) distribution.CoefficientOfVariation;
         }
     }
 }
